requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
